+++ nifti2bids/io.py
@@ -1,7 +1,8 @@
 """Module for input/output operations."""
 
-import shutil
+import shutil, re
 from pathlib import Path
+from typing import Optional
 
 import nibabel as nib
 from numpy.typing import NDArray
@@ -35,7 +36,12 @@ def load_nifti(
     return nifti_img
 
 
-def compress_image(nifti_file: str | Path, remove_src_file: bool = False) -> None:
+def compress_image(
+    nifti_file: str | Path,
+    dst_dir: Optional[str | Path] = None,
+    remove_src_file: bool = False,
+    return_dst_file: bool = False,
+) -> Path | None:
     """
     Compresses a ".nii" image to a ".nii.gz" image.
 
@@ -44,23 +50,40 @@ def compress_image(nifti_file: str | Path, remove_src_file: bool = False) -> Non
     nifti_file: :obj:`str` or :obj:`Path`
         Path to the NIfTI image.
 
-    remove_src_file: :obj:`bool`
+    dst_dir: :obj:`str` or :obj:`Path`, default=None
+        Destination directory for the NIfTI image. If None, image is saved in the
+        source directory.
+
+    remove_src_file: :obj:`bool`, default=False
         Deletes the original source image file.
+
+    return_dst_file: :obj:`bool`, default=False
+        Return the path to the compressed file.
 
     Returns
     -------
-    None
+    Path or None
+        Path to compressed file if ``return_dst_file`` is True else None.
     """
     img = nib.load(nifti_file)
-    nib.save(img, str(nifti_file).replace(".nii", ".nii.gz"))
+
+    nifti_file = Path(nifti_file)
+    dst_dir = Path(dst_dir) if dst_dir else nifti_file.parent
+
+    dst_file = dst_dir / str(nifti_file.name).replace(".nii", ".nii.gz")
+    nib.save(img, dst_file)
 
     if remove_src_file:
-        Path(nifti_file).unlink()
+        nifti_file.unlink()
+
+    return dst_file if return_dst_file else None
 
 
-def glob_contents(src_dir: str | Path, pattern: str) -> list[Path]:
+def regex_glob(
+    src_dir: str | Path, pattern: str, recursive: bool = False
+) -> list[Path]:
     """
-    Use glob to get contents with specific patterns.
+    Use regex to get content in the source directory with specific patterns.
 
     Parameters
     ----------
@@ -68,14 +91,28 @@ def glob_contents(src_dir: str | Path, pattern: str) -> list[Path]:
         The source directory.
 
     pattern: :obj:`str`
-        The wildcard pattern.
+        The regex pattern.
+
+    recursive: :obj:`bool`, default=False
+        If True, regex pattern is applied to content in the top-level directory
+        (i.e., sub-101.log) and nested directories (i.e. logs/sub-101.log). If
+        False, regex pattern is only applied to content in the top-level directory.
+
 
     Returns
     -------
     list[Path]
-        List of contents with the pattern specified by ``pattern``.
+        List of contents filtered by the regex pattern specified by ``pattern``.
+
+    Example
+    -------
+    >>> from nifti2bids.io import regex_glob
+    >>> # Get any file ending in pdf or txt
+    >>> regex_glob(r"path/to/directory", pattern=r"^.*.(pdf|txt)$")
     """
-    return list(Path(src_dir).glob(f"{pattern}"))
+    all_contents = Path(src_dir).rglob("*") if recursive else Path(src_dir).glob("*")
+
+    return [path for path in all_contents if re.compile(pattern).match(path.name)]
 
 
 def get_nifti_header(
@@ -115,7 +152,7 @@ def get_nifti_affine(nifti_file_or_img: str | Path | nib.nifti1.Nifti1Image) -> 
 
 
 def _copy_file(
-    src_file: str | Path, dst_file: str | Path, remove_src_file: bool
+    src_file: str | Path, dst_file: str | Path, remove_src_file: bool = False
 ) -> None:
     """
     Copy a file and optionally remove the source file.
@@ -128,12 +165,8 @@ def _copy_file(
     dst_file: :obj:`str` or :obj:`Path`
         The new destination file.
 
-    remove_src_file: :obj:`bool`
+    remove_src_file: :obj:`bool`, default=False
         Delete the source file if True.
-
-    Returns
-    -------
-    None
     """
     shutil.copy(src_file, dst_file)